chico_bot/requirements.txt
# ChicoBot - Dépendances Python

# 🤖 Framework Telegram
aiogram>=3.4.1
aiohttp>=3.9.0

# 🗄️ Base de Données
sqlalchemy>=2.0.23
aiosqlite>=0.19.0
alembic>=1.13.0

# 🔐 Sécurité et Cryptographie
cryptography>=41.0.7
pycryptodome>=3.19.0
passlib>=1.7.4
python-jose>=3.3.0

# 🤖 Services IA (Chico Personality)
openai>=1.3.7
google-generativeai>=0.3.2
transformers>=4.35.2
torch>=2.1.1

# 📊 Data Science et Trading
pandas>=2.1.3
numpy>=1.24.3
scipy>=1.11.4
scikit-learn>=1.3.2
statsmodels>=0.14.0
psutil>=5.9.6

# 🌐 Web Scraping et APIs
beautifulsoup4>=4.12.2
requests>=2.31.0
selenium>=4.15.2
selenium-wire>=5.1.0

# ⏰ Tâches et Cache (Multitâche)
redis>=5.0.1
celery>=5.3.4
apscheduler>=3.10.4

# 📝 Configuration et Validation
pydantic>=2.5.0
python-dotenv>=1.0.0

# 📊 Logging et Monitoring
loguru>=0.7.2
prometheus-client>=0.19.0

# 🧪 Tests
pytest>=7.4.3
pytest-asyncio>=0.21.1
pytest-mock>=3.12.0

# 🔧 Développement
black>=23.11.0
flake8>=6.1.0
mypy>=1.7.1

# 📦 Utilitaires
tqdm>=4.66.1
colorama>=0.4.6
python-dateutil>=2.8.2
pytz>=2023.3

# 🚀 Production
gunicorn>=21.2.0
uvicorn>=0.24.0
